requirements.txt
art==6.3
aiogram==3.10.0
tenacity==8.5.0
python-dotenv==1.0.0
PyYAML==6.0.1
pyuseragents==1.0.5
curl_cffi==0.7.1
captchatools
Pyrogram~=2.0.106
paho-mqtt==2.1.0
imap-tools==1.6.0
cryptography==42.0.5
datetime==5.5
better-proxy==1.1.5
httpx==0.27.2